Important update.

+++ scripts/Modding/SetupFilePatching.ps1
@@ -2,11 +2,16 @@
 
 . $PathScriptsRoot/Modding/Modding
 
- $ModPrefixDirectories = [System.Collections.ArrayList]@("ExampleMod")
- $ModPrefixDirectories.Add("HM") > $null
-#  $ModPrefixDirectories.Add("DayZEditor") > $null
-#  $ModPrefixDirectories.Add("JM") > $null
-#  $ModPrefixDirectories.Add("DayZExpansion") > $null
+$ModPrefixDirectories = [System.Collections.ArrayList]@("ExampleMod")
+
+# $ModPrefixDirectories.Add("ExampleMod") > $null
+$ModPrefixDirectories.Add("HM") > $null
+# $ModPrefixDirectories.Add("TWS") > $null
+# $ModPrefixDirectories.Add("TheWalkingSurvivorsMod") > $null
+# $ModPrefixDirectories.Add("TWSMod") > $null
+
+# $ModPrefixDirectories.Add("JM") > $null
+# $ModPrefixDirectories.Add("DayZExpansion") > $null
 
 $ModPrefixDirectoriesSize = $ModPrefixDirectories.Count
 
@@ -14,12 +19,17 @@ foreach ($prefix in $ModPrefixDirectories) {
   # "$prefix"
 }
 
-# rmdir "$PathDayZWorkbench\Addons\"
+
+if (Test-Path -Path "$PathDayZWorkbench\Addons\" -PathType Container) {
+  cmd /c rmdir "$PathDayZWorkbench\Addons\"
+}
 cmd /c mklink /J "$PathDayZWorkbench\Addons\" "$PathClient\Addons\"
 
 foreach ($prefix in $ModPrefixDirectories) {
   # Server
-  cmd /c rmdir "$PathServer\$prefix\"
+  if (Test-Path -Path "$PathServer\$prefix\" -PathType Container) {
+    cmd /c rmdir "$PathServer\$prefix\"
+  }
   cmd /c mklink /J "$PathServer\$prefix\" "$PathWorkDrive\$prefix\"
 
   # cmd /c rmdir "$PathDayZServer\$prefix\"
@@ -27,9 +37,10 @@ foreach ($prefix in $ModPrefixDirectories) {
 
   # Client
   # cmd /c mklink /J "%ClientDirectory%\!ModPrefixDirectories[%%n]!\" "%WorkDriveRoot%!ModPrefixDirectories[%%n]!\"\
-  cmd /c rmdir "$PathClient\$prefix\"
+  if (Test-Path -Path "$PathClient\$prefix\" -PathType Container) {
+    cmd /c rmdir "$PathClient\$prefix\"
+  }
   cmd /c mklink /J "$PathClient\$prefix\" "$PathWorkDrive\$prefix\"
-
   # cmd /c rmdir "$PathDayZClient\$prefix\"
   # cmd /c mklink /J "$PathDayZClient\$prefix\" "$PathWorkDrive\$prefix\"
 }
